requirements.txt
# ============================================
# SITNOVA - Dependencias del Agente Principal
# ============================================

# ============================================
# CORE
# ============================================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# FASTAPI
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# ============================================
# LANGCHAIN & LANGGRAPH
# ============================================
langchain==0.1.0
langchain-core==0.1.10
langchain-anthropic==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
langsmith==0.0.77

# ============================================
# SUPABASE
# ============================================
supabase==2.3.0
postgrest==0.13.0

# ============================================
# REDIS
# ============================================
redis==5.0.1
redis-om==0.2.1

# ============================================
# HTTP CLIENTS
# ============================================
httpx==0.26.0
requests==2.31.0

# ============================================
# UTILITIES
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# ============================================
# MONITORING & LOGGING
# ============================================
loguru==0.7.2
prometheus-client==0.19.0

# ============================================
# TESTING
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0  # para TestClient
